Move GlobalVariableName to components
