+++ packages/studio-base/src/components/GlobalVariableName.tsx
